refactor: Move getCustomToggles to parseKeyValue
