+++ src/ts/process/prompt.ts
@@ -64,20 +64,3 @@ export async function tokenizePreset(prompts:PromptItem[], consti:boolean = fals
     }
     return total
 }
-
-export function getCustomToggles(template:string){
-    if(!template){
-        return []
-    }
-
-    const keyValue:[string, string][] = []
-
-    for(const line of template.split('\n')){
-        const [key, value] = line.split('=')
-        if(key && value){
-            keyValue.push([key, value])
-        }
-    }
-
-    return keyValue
-}
